requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
speechrecognition
gtts
playsound